requirements.txt
greenlet==1.1.3.post0
Pillow==9.2.0
SQLAlchemy==1.4.42
